* Przywr?cenie starej pozycji elementu w menu Program (z wersji 1.0.0.0). * Ma?e poprawki w kodzie. * Wtyczka sprawdza przy ?adowaniu czy w katalogu aktywnej kompozycji w folderze Icons znajduje si? plik "AQQRestarter.png", je?eli tak ustawia go jako grafik? przycisku.

+++ AQQRestarter.cpp
@@ -3,7 +3,7 @@
 #include <windows.h>
 #pragma hdrstop    
 #pragma argsused
-#include "Aqq.h"
+#include "Aqq.h"  
 #include <memory> 
 #include "ikonka.rh"
 //---------------------------------------------------------------------------
@@ -15,6 +15,14 @@ int WINAPI DllEntryPoint(HINSTANCE hinst, unsigned long reason, void* lpReserved
   hInstance = hinst;
   return 1;
 }
+//---------------------------------------------------------------------------
+
+wchar_t *AnsiTowchar_t(AnsiString Str) //zamiana AnsiString->wchar_t*
+{
+  wchar_t *str = new wchar_t[Str.WideCharBufSize()];
+  return Str.WideChar(str, Str.WideCharBufSize());
+}
+//---------------------------------------------------------------------------
 
 //Utworzenie obiektow do struktur
 PluginAction TPluginActionMenu;
@@ -32,13 +40,14 @@ int __stdcall AqqReStartService (WPARAM, LPARAM)
 
   return 0;
 }
+//---------------------------------------------------------------------------
 
 //Program
 extern "C"  __declspec(dllexport) PluginInfo* __stdcall AQQPluginInfo(DWORD AQQVersion)
 {
   TPluginInfo.cbSize = sizeof(PluginInfo);
   TPluginInfo.ShortName = (wchar_t*)L"AQQ Restarter";
-  TPluginInfo.Version = PLUGIN_MAKE_VERSION(1,0,0,0);
+  TPluginInfo.Version = PLUGIN_MAKE_VERSION(1,0,1,2);
   TPluginInfo.Description = (wchar_t *)L"Szybki restart AQQ z pozycji menu";
   TPluginInfo.Author = (wchar_t *)L"Krzysztof Grochocki (Beherit)";
   TPluginInfo.AuthorMail = (wchar_t *)L"beherit666@vp.pl";
@@ -47,20 +56,21 @@ extern "C"  __declspec(dllexport) PluginInfo* __stdcall AQQPluginInfo(DWORD AQQV
 
   return &TPluginInfo;
 }
+//---------------------------------------------------------------------------
 
 void PrzypiszSkrotMenu()
 {
   TPluginActionMenu.cbSize = sizeof(PluginAction);
   TPluginActionMenu.pszName = (wchar_t*)L"AQQRestarterPrzypiszSkrotMenu";
   TPluginActionMenu.pszCaption = (wchar_t*) L"Zrestartuj AQQ";
-  TPluginActionMenu.Position = 9;
+  TPluginActionMenu.Position = 12;
   TPluginActionMenu.IconIndex = plugin_icon_idx;
   TPluginActionMenu.pszService = (wchar_t*) L"serwis_aqqrestart";
   TPluginActionMenu.pszPopupName = (wchar_t*) L"muProgram";
 
   TPluginLink.CallService(AQQ_CONTROLS_CREATEPOPUPMENUITEM,0,(LPARAM)(&TPluginActionMenu));
-  TPluginLink.CreateServiceFunction(L"serwis_aqqrestart",AqqReStartService);
 }
+//---------------------------------------------------------------------------
 
 void PrzypiszSkrotMakra()
 {
@@ -73,8 +83,8 @@ void PrzypiszSkrotMakra()
   TPluginActionMakra.pszPopupName = (wchar_t*) L"popMacros";
 
   TPluginLink.CallService(AQQ_CONTROLS_CREATEPOPUPMENUITEM,0,(LPARAM)(&TPluginActionMakra));
-  TPluginLink.CreateServiceFunction(L"serwis_aqqrestart",AqqReStartService);
 }
+//---------------------------------------------------------------------------
 
 //Do wypakowywania RES
 void ExtractExe(unsigned short ID, AnsiString FileName)
@@ -97,37 +107,55 @@ void ExtractExe(unsigned short ID, AnsiString FileName)
   Ms->SaveToFile(FileName);
   Ms->Free();
 }
+//---------------------------------------------------------------------------
 
 extern "C" int __declspec(dllexport) __stdcall Load(PluginLink *Link)
 {
   TPluginLink = *Link;
 
-  //Katalog wtyczki
-  AnsiString PluginPath = (wchar_t*)(TPluginLink.CallService(AQQ_FUNCTION_GETPLUGINUSERDIR,(WPARAM)(hInstance),0));
+  //Katalog aktywnej kompozycji
+  AnsiString PluginPath = (wchar_t*)(TPluginLink.CallService(AQQ_FUNCTION_GETTHEMEDIR,(WPARAM)(hInstance),0));
   PluginPath = StringReplace(PluginPath, "\\", "\\\\", TReplaceFlags() << rfReplaceAll);
 
-  if(!FileExists(PluginPath + "\\\\AQQRestarter.png"))
+  if(!FileExists(PluginPath + "\\\\Icons\\\\AQQRestarter.png"))
   {
-    //Wypakowanie ikon
-    ExtractExe(ID_PNG,"AQQRestarter.png");
-    //Przypisanie ikony
-    wchar_t* plugin_icon_path = L"AQQRestarter.png";
-    plugin_icon_idx=TPluginLink.CallService(AQQ_ICONS_LOADPNGICON,0, (LPARAM)(plugin_icon_path));
-    //Usuniecie ikony
-    DeleteFile("AQQRestarter.png");
+    //Katalog prywatny wtyczel
+    PluginPath = (wchar_t*)(TPluginLink.CallService(AQQ_FUNCTION_GETPLUGINUSERDIR,(WPARAM)(hInstance),0));
+    PluginPath = StringReplace(PluginPath, "\\", "\\\\", TReplaceFlags() << rfReplaceAll);
+
+    if(!FileExists(PluginPath + "\\\\AQQRestarter.png"))
+    {
+      //Wypakowanie ikon
+      ExtractExe(ID_PNG,PluginPath + "\\\\AQQRestarter.png");
+      //Przypisanie ikony
+      wchar_t* plugin_icon_path = AnsiTowchar_t(PluginPath + "\\\\AQQRestarter.png");
+      plugin_icon_idx=TPluginLink.CallService(AQQ_ICONS_LOADPNGICON,0, (LPARAM)(plugin_icon_path));
+      //Usuniecie ikony
+      DeleteFile("AQQRestarter.png");
+    }
+    else
+    {
+      //Przypisanie ikony
+      wchar_t* plugin_icon_path = AnsiTowchar_t(PluginPath + "\\\\AQQRestarter.png");
+      plugin_icon_idx=TPluginLink.CallService(AQQ_ICONS_LOADPNGICON,0, (LPARAM)(plugin_icon_path));
+    }
   }
   else
   {
     //Przypisanie ikony
-    wchar_t* plugin_icon_path = L"AQQRestarter.png";
+    wchar_t* plugin_icon_path = AnsiTowchar_t(PluginPath + "\\\\Icons\\\\AQQRestarter.png");
     plugin_icon_idx=TPluginLink.CallService(AQQ_ICONS_LOADPNGICON,0, (LPARAM)(plugin_icon_path));
   }
 
+  //Utworzenie serwisu restartu
+  TPluginLink.CreateServiceFunction(L"serwis_aqqrestart",AqqReStartService);
+  //Przypisanie skrutow
   PrzypiszSkrotMenu();
   PrzypiszSkrotMakra();
 
   return 0;
 }
+//---------------------------------------------------------------------------
 
 extern "C" int __declspec(dllexport) __stdcall Unload()
 {
@@ -138,3 +166,4 @@ extern "C" int __declspec(dllexport) __stdcall Unload()
  
   return 0;
 }
+//---------------------------------------------------------------------------
